NIFI-11385 Moved JMX Metrics method to System Diagnostics Resource
