+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/java/org/apache/nifi/web/NiFiWebApiResourceConfig.java
@@ -103,7 +103,6 @@ public class NiFiWebApiResourceConfig extends ResourceConfig {
         register(ctx.getBean("versionsResource"));
         register(ctx.getBean("parameterContextResource"));
         register(ctx.getBean("parameterProviderResource"));
-        register(ctx.getBean("jmxMetricsResource"));
 
         // exception mappers
         register(AccessDeniedExceptionMapper.class);
