+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-api/src/main/resources/nifi-web-api-context.xml
@@ -608,6 +608,7 @@
         <property name="requestReplicator" ref="requestReplicator" />
         <property name="authorizer" ref="authorizer"/>
         <property name="flowController" ref="flowController" />
+        <property name="jmxMetricsService" ref="jmxMetricsService" />
     </bean>
     <bean id="accessResource" class="org.apache.nifi.web.api.AccessResource" scope="singleton">
         <property name="logoutRequestManager" ref="logoutRequestManager" />
@@ -655,14 +656,8 @@
         <property name="clusterComponentLifecycle" ref="clusterComponentLifecycle" />
         <property name="localComponentLifecycle" ref="localComponentLifecycle" />
     </bean>
-    <bean id="jmxMetricsResource" class="org.apache.nifi.web.api.JmxMetricsResource" scope="singleton">
-        <property name="serviceFacade" ref="serviceFacade"/>
-        <property name="authorizer" ref="authorizer"/>
-        <property name="jmxMetricsService" ref="jmxMetricsService"/>
-    </bean>
     <bean id="jmxMetricsService" class="org.apache.nifi.web.api.metrics.jmx.StandardJmxMetricsService">
         <property name="properties" ref="nifiProperties"/>
-        <property name="metricsResultConverter" ref="metricsResultConverter"/>
         <property name="metricsCollector" ref="metricsCollector"/>
     </bean>
     <bean id="metricsResultConverter" class="org.apache.nifi.web.api.metrics.jmx.JmxMetricsResultConverter">
